src/echo/requirements.txt
Flask==2.2.2
PyYAML==5.3.1
requests==2.25.1
Flask-RESTful==0.3.9
PyJWT==1.7.1
jwcrypto==1.4.2